pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dpo"
version = "0.1.0"
description = "Data Profiling Orchestrator - Automate Databricks Data Profiling at scale"
readme = "README.md"
license = {text = "GPL-3.0-only"}
requires-python = ">=3.11"
authors = [
    {name = "Databricks Solution Accelerators"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "databricks-sdk>=0.77.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "tenacity>=8.0",
    "tabulate>=0.9.0",
    "pytest-cov>=7.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-mock",
    "black",
    "ruff",
]

[project.urls]
"Homepage" = "https://github.com/databricks/solution-accelerators"
"Bug Tracker" = "https://github.com/databricks/solution-accelerators/issues"

[tool.setuptools.packages.find]
where = ["src"]
include = ["dpo*"]

[tool.black]
line-length = 120
target-version = ['py311', 'py312']

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]
